switch to nightly autodetection
