+++ Cargo.toml
@@ -22,20 +22,13 @@ edition = "2018"
 azure-devops = { project = "alecmocatta/serde_closure", pipeline = "tests", build = "10" }
 maintenance = { status = "actively-developed" }
 
-[features]
-nightly = []
-
-[package.metadata.docs.rs]
-features = ["nightly"]
-
 [dependencies]
 serde_closure_derive = { version = "=0.3.0", path = "serde_closure_derive" }
 serde = { version = "1.0", features = ["derive"] }
 
+[build-dependencies]
+rustversion = "1.0"
+
 [dev-dependencies]
 serde_json = "1.0"
 bincode = "1.0"
-
-[[test]]
-name = "test"
-required-features = ["nightly"]
