+++ azure-pipelines.yml
@@ -13,11 +13,11 @@ jobs:
   parameters:
     endpoint: alecmocatta
     default:
-      rust_toolchain: nightly
-      rust_lint_toolchain: nightly-2020-07-09
+      rust_toolchain: beta nightly
+      rust_lint_toolchain: nightly-2020-07-12
       rust_flags: ''
-      rust_features_clippy: ';nightly'
-      rust_features: 'nightly'
+      rust_features_clippy: ''
+      rust_features: ''
       rust_target_check: ''
       rust_target_build: ''
       rust_target_run: ''
